fix: repair email service with proper SMTP and Resend support
- Fix sendVerificationCodeEmail function with Resend and SMTP fallback
- Add proper error handling and email service validation
- Update register route to check email send status
- Add debugging logs for email configuration
- Clean up corrupted email.ts file
- Support both Resend API and Gmail SMTP configurations
